worker/content-moderation/app/requirements.txt
pika==1.3.2
pyClamd==0.4.0
boto3==1.34.7
requests==2.31.0
tensorflow==2.15.1
pillow==10.3.0
transformers==4.40.1
python-dotenv==1.0.0